+++ src/resources/users/user.router.ts
@@ -10,7 +10,7 @@ import {
 // import { IUser } from '../intefases';
 const routerUser = Router();
 
-routerUser.route('/').get(async (req: Request, res: Response) => {
+routerUser.route('/').get(async (_req: Request, res: Response) => {
   const users = await getAllServis();
   res.json(users.map(User.toResponse));
 });
@@ -21,27 +21,36 @@ routerUser.route('/').post(async (req: Request, res: Response) => {
   res.status(201).json(User.toResponse(usersPost));
 });
 
-routerUser.route('/:id').get(async (req: Request, res: Response) => {
-  const idNumber = req.params.id;
-  const usersid = await getIDServis(idNumber);
-  if (!usersid) {
-    res.status(404).json();
-  } else {
-    res.status(200).json(User.toResponse(usersid));
+routerUser.route('/:userId').get(async (req: Request, res: Response) => {
+  const idNumber = req.params['userId'];
+  if (idNumber !== undefined) {
+    const usersid = await getIDServis(idNumber);
+    if (!usersid) {
+      res.status(404).json();
+    } else {
+      res.status(200).json(User.toResponse(usersid));
+    }
   }
 });
 
 routerUser.route('/:id').put(async (req: Request, res: Response) => {
-  const idNumberPut = req.params.id;
+  const idNumberPut = req.params['id'];
   const createUserPut = new User(req.body);
-  const usersPut = await putUserServis(idNumberPut, createUserPut);
-  res.status(200).json(User.toResponse(usersPut));
+  if (idNumberPut !== undefined) {
+    const usersPut = await putUserServis(createUserPut, idNumberPut);
+    if (usersPut !== undefined) {
+      res.status(200).json(User.toResponse(usersPut));
+    } else {
+      res.status(404).json();
+    }
+  }
 });
 
 routerUser.route('/:id').delete(async (req: Request, res: Response) => {
-  const idNumberDelete = req.params.id;
+  const idNumberDelete = req.params['id'];
+  if (!idNumberDelete) return res.status(200).json({});
   const usersDelete = await deleteUserServis(idNumberDelete);
-  res.status(200).json(User.toResponse(usersDelete));
+  return res.status(200).json(User.toResponse(usersDelete));
 });
 
 export { routerUser };
